# api/openapi.yaml
openapi: 3.1.0
info:
  title: uspace-cisp national CIS publication API
  version: 0.1.0
  description: |
    The Common Information Service of the Georgian U-space
    system-of-systems (Reg. (EU) 2021/664 Art. 5). This file is the
    contract: an endpoint that is not in it does not exist, and changes
    within /v1 are additive only. Every error is
    application/problem+json in the shape of the Problem schema.

    Skeleton (WP-0): health, readiness and the status stub. Every other
    work package adds its paths under its tag; see api/README.md.
    WP-2: the signing JWKS.
servers:
  - url: /
...
paths:
...
  /v1/status:
    get:
      operationId: getStatus
      tags: [status]
      summary: Service status (stub)
      description: |
        Per dataset the current version, per publisher the heartbeat and
        staleness, and the degraded dependencies with their since-times.
        Not implemented yet: answers 501 not_implemented.
      security:
        - ecosystemToken: []
        - consoleSession: []
      responses:
        "200":
          description: The service status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
        "501":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /.well-known/jwks.json:
    get:
      operationId: getJWKS
      tags: [keys]
      summary: The CISP's signing keys (JWKS)
      description: |
        The public keys of the CISP's signing key ring (RFC 7517): the key
        that signs, and during a rotation the previous one, each with
        kid, alg RS256 and use sig. Webhook deliveries and signed
        snapshots are verified with it. No authentication; cacheable for
        an hour; If-None-Match with the current ETag answers 304.
      security: []
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: The key set.
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWKS"
        "304":
          description: The key set has not changed.
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    ecosystemToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 machine token from the authority's token service; aud is this host.
    consoleSession:
      type: http
      scheme: bearer
      description: Console session JWT forwarded by the web BFF (scope "session").
  responses:
    Problem:
      description: An error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: |
        The ecosystem-wide error body (RFC 9457), the same shape as
        uspace-lab schemas/common/problem/v1.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: https://schemas.uspace.ge/problems/<slug>
          examples: ["https://schemas.uspace.ge/problems/not_implemented"]
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          description: The request path, or the request id.
        errors:
          type: array
          description: Field problems, at most 100; truncated says when more existed.
          maxItems: 100
          items:
            $ref: "#/components/schemas/FieldProblem"
        truncated:
          type: boolean
    FieldProblem:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
          description: The JSON path or parameter name.
        reason:
          type: string
    JWKS:
      type: object
      description: A JSON Web Key Set (RFC 7517 section 5) of RSA signing keys.
      required: [keys]
      properties:
        keys:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/JWK"
    JWK:
      type: object
      description: One RSA public key (RFC 7517, RFC 7518 section 6.3.1).
      required: [kty, kid, use, alg, n, e]
      properties:
        kty:
          type: string
          enum: [RSA]
        kid:
          type: string
        use:
          type: string
          enum: [sig]
        alg:
          type: string
          enum: [RS256]
        n:
          type: string
          description: The modulus, base64url without padding.
        e:
          type: string
          description: The exponent, base64url without padding.
...
    Status:
      type: object
      description: Stub of the status document (docs/PLAN.md section 6.3).
      required: [now, datasets, publishers, degraded]
      properties:
        now:
          type: string
          format: date-time
        datasets:
          type: array
          items:
            type: object
            required: [dataset, current_version]
            properties:
              dataset:
                type: string
              current_version:
                type: integer
                format: int64
                minimum: 0
              updated_at:
                type: string
                format: date-time
        publishers:
          type: array
          items:
            type: object
            required: [client_id, stale]
            properties:
              client_id:
                type: string
              last_heartbeat_at:
                type: string
                format: date-time
              stale:
                type: boolean
        degraded:
          type: array
          items:
            type: object
            required: [component, since]
            properties:
              component:
                type: string
                enum: [database, nats, jwks]
              since:
                type: string
                format: date-time