# internal/httpapi/gen/api.gen.go
//go:build go1.22

// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for HealthStatus.
const (
	Ok HealthStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatus enum.
func (e HealthStatus) Valid() bool {
	switch e {
	case Ok:
		return true
	default:
		return false
	}
}

// Defines values for JWKAlg.
const (
	RS256 JWKAlg = "RS256"
)

// Valid indicates whether the value is a known member of the JWKAlg enum.
func (e JWKAlg) Valid() bool {
	switch e {
	case RS256:
		return true
	default:
		return false
	}
}

// Defines values for JWKKty.
const (
	RSA JWKKty = "RSA"
)

// Valid indicates whether the value is a known member of the JWKKty enum.
func (e JWKKty) Valid() bool {
	switch e {
	case RSA:
		return true
	default:
		return false
	}
}

// Defines values for JWKUse.
const (
	Sig JWKUse = "sig"
)

// Valid indicates whether the value is a known member of the JWKUse enum.
func (e JWKUse) Valid() bool {
	switch e {
	case Sig:
		return true
	default:
		return false
	}
}

// Defines values for ReadinessStatus.
const (
	NotReady ReadinessStatus = "not_ready"
	Ready    ReadinessStatus = "ready"
)

// Valid indicates whether the value is a known member of the ReadinessStatus enum.
func (e ReadinessStatus) Valid() bool {
	switch e {
	case NotReady:
		return true
	case Ready:
		return true
	default:
		return false
	}
}

// Defines values for StatusDegradedComponent.
const (
	Database StatusDegradedComponent = "database"
	Jwks     StatusDegradedComponent = "jwks"
	Nats     StatusDegradedComponent = "nats"
)

// Valid indicates whether the value is a known member of the StatusDegradedComponent enum.
func (e StatusDegradedComponent) Valid() bool {
	switch e {
	case Database:
		return true
	case Jwks:
		return true
	case Nats:
		return true
	default:
		return false
	}
}

// FieldProblem defines model for FieldProblem.
type FieldProblem struct {
	// Field The JSON path or parameter name.
	Field  string `json:"field"`
	Reason string `json:"reason"`
}

// Health defines model for Health.
type Health struct {
	Status HealthStatus `json:"status"`
}

// HealthStatus defines model for Health.Status.
type HealthStatus string

// JWK One RSA public key (RFC 7517, RFC 7518 section 6.3.1).
type JWK struct {
	Alg JWKAlg `json:"alg"`

	// E The exponent, base64url without padding.
	E   string `json:"e"`
	Kid string `json:"kid"`
	Kty JWKKty `json:"kty"`

	// N The modulus, base64url without padding.
	N   string `json:"n"`
	Use JWKUse `json:"use"`
}

// JWKAlg defines model for JWK.Alg.
type JWKAlg string

// JWKKty defines model for JWK.Kty.
type JWKKty string

// JWKUse defines model for JWK.Use.
type JWKUse string

// JWKS A JSON Web Key Set (RFC 7517 section 5) of RSA signing keys.
type JWKS struct {
	Keys []JWK `json:"keys"`
}

// Problem The ecosystem-wide error body (RFC 9457), the same shape as
// uspace-lab schemas/common/problem/v1.
type Problem struct {
	Detail *string `json:"detail,omitempty"`

	// Errors Field problems, at most 100; truncated says when more existed.
	Errors *[]FieldProblem `json:"errors,omitempty"`

	// Instance The request path, or the request id.
	Instance  *string `json:"instance,omitempty"`
	Status    int     `json:"status"`
	Title     string  `json:"title"`
	Truncated *bool   `json:"truncated,omitempty"`

	// Type https://schemas.uspace.ge/problems/<slug>
	//
	// Examples: https://schemas.uspace.ge/problems/not_implemented
	Type string `json:"type"`
}

// Readiness defines model for Readiness.
type Readiness struct {
	Checks struct {
		// Database ok, not configured, or unreachable with the reason.
		Database string `json:"database"`

		// Migrations ok, or pending with what is pending.
		Migrations string `json:"migrations"`

		// Nats ok, not configured, or disconnected.
		Nats string `json:"nats"`
	} `json:"checks"`
	Status ReadinessStatus `json:"status"`
}

// ReadinessStatus defines model for Readiness.Status.
type ReadinessStatus string

// Status Stub of the status document (docs/PLAN.md section 6.3).
type Status struct {
	Datasets []struct {
		CurrentVersion int64      `json:"current_version"`
		Dataset        string     `json:"dataset"`
		UpdatedAt      *time.Time `json:"updated_at,omitempty"`
	} `json:"datasets"`
	Degraded []struct {
		Component StatusDegradedComponent `json:"component"`
		Since     time.Time               `json:"since"`
	} `json:"degraded"`
	Now        time.Time `json:"now"`
	Publishers []struct {
		ClientId        string     `json:"client_id"`
		LastHeartbeatAt *time.Time `json:"last_heartbeat_at,omitempty"`
		Stale           bool       `json:"stale"`
	} `json:"publishers"`
}

// StatusDegradedComponent defines model for Status.Degraded.Component.
type StatusDegradedComponent string

// GetJWKSParams defines parameters for GetJWKS.
type GetJWKSParams struct {
	IfNoneMatch *string `json:"If-None-Match,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetJWKS The CISP's signing keys (JWKS)
	// (GET /.well-known/jwks.json)
	GetJWKS(w http.ResponseWriter, r *http.Request, params GetJWKSParams)
	// GetHealthz Liveness
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// GetReadyz Readiness
	// (GET /readyz)
	GetReadyz(w http.ResponseWriter, r *http.Request)
	// GetStatus Service status (stub)
	// (GET /v1/status)
	GetStatus(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetJWKS operation middleware
func (siw *ServerInterfaceWrapper) GetJWKS(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetJWKSParams

	headers := r.Header

	// ------------- Optional header parameter "If-None-Match" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("If-None-Match")]; found {
		var IfNoneMatch string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "If-None-Match", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "If-None-Match", valueList[0], &IfNoneMatch, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "If-None-Match", Err: err})
			return
		}

		params.IfNoneMatch = &IfNoneMatch

	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetJWKS(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetReadyz operation middleware
func (siw *ServerInterfaceWrapper) GetReadyz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetReadyz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetStatus operation middleware
func (siw *ServerInterfaceWrapper) GetStatus(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetStatus(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/healthz", wrapper.GetHealthz)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/readyz", wrapper.GetReadyz)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/status", wrapper.GetStatus)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/.well-known/jwks.json", wrapper.GetJWKS)

	return m
}

type ProblemApplicationProblemPlusJSONResponse Problem

type GetJWKSRequestObject struct {
	Params GetJWKSParams
}

type GetJWKSResponseObject interface {
	VisitGetJWKSResponse(w http.ResponseWriter) error
}

type GetJWKS200ResponseHeaders struct {
	CacheControl *string
	ETag         *string
}

type GetJWKS200JSONResponse struct {
	Body    JWKS
	Headers GetJWKS200ResponseHeaders
}

func (response GetJWKS200JSONResponse) VisitGetJWKSResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.CacheControl != nil {
		w.Header().Set("Cache-Control", fmt.Sprint(*response.Headers.CacheControl))
	}
	if response.Headers.ETag != nil {
		w.Header().Set("ETag", fmt.Sprint(*response.Headers.ETag))
	}
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetJWKS304ResponseHeaders struct {
	CacheControl *string
	ETag         *string
}

type GetJWKS304Response struct {
	Headers GetJWKS304ResponseHeaders
}

func (response GetJWKS304Response) VisitGetJWKSResponse(w http.ResponseWriter) error {
	if response.Headers.CacheControl != nil {
		w.Header().Set("Cache-Control", fmt.Sprint(*response.Headers.CacheControl))
	}
	if response.Headers.ETag != nil {
		w.Header().Set("ETag", fmt.Sprint(*response.Headers.ETag))
	}
	w.WriteHeader(304)
	return nil
}

type GetJWKS503ApplicationProblemPlusJSONResponse struct {
	ProblemApplicationProblemPlusJSONResponse
}

func (response GetJWKS503ApplicationProblemPlusJSONResponse) VisitGetJWKSResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

type GetJWKSdefaultApplicationProblemPlusJSONResponse struct {
	Body       Problem
	StatusCode int
}

func (response GetJWKSdefaultApplicationProblemPlusJSONResponse) VisitGetJWKSResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetHealthzRequestObject struct {
}

type GetHealthzResponseObject interface {
	VisitGetHealthzResponse(w http.ResponseWriter) error
}

type GetHealthz200JSONResponse Health

func (response GetHealthz200JSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetHealthzdefaultApplicationProblemPlusJSONResponse struct {
	Body       Problem
	StatusCode int
}

func (response GetHealthzdefaultApplicationProblemPlusJSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyzRequestObject struct {
}

type GetReadyzResponseObject interface {
	VisitGetReadyzResponse(w http.ResponseWriter) error
}

type GetReadyz200JSONResponse Readiness

func (response GetReadyz200JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyz503JSONResponse Readiness

func (response GetReadyz503JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyzdefaultApplicationProblemPlusJSONResponse struct {
	Body       Problem
	StatusCode int
}

func (response GetReadyzdefaultApplicationProblemPlusJSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetStatusRequestObject struct {
}

type GetStatusResponseObject interface {
	VisitGetStatusResponse(w http.ResponseWriter) error
}

type GetStatus200JSONResponse Status

func (response GetStatus200JSONResponse) VisitGetStatusResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetStatus501ApplicationProblemPlusJSONResponse struct {
	ProblemApplicationProblemPlusJSONResponse
}

func (response GetStatus501ApplicationProblemPlusJSONResponse) VisitGetStatusResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(501)
	_, err := buf.WriteTo(w)
	return err
}

type GetStatusdefaultApplicationProblemPlusJSONResponse struct {
	Body       Problem
	StatusCode int
}

func (response GetStatusdefaultApplicationProblemPlusJSONResponse) VisitGetStatusResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetJWKS The CISP's signing keys (JWKS)
	// (GET /.well-known/jwks.json)
	GetJWKS(ctx context.Context, request GetJWKSRequestObject) (GetJWKSResponseObject, error)
	// GetHealthz Liveness
	// (GET /healthz)
	GetHealthz(ctx context.Context, request GetHealthzRequestObject) (GetHealthzResponseObject, error)
	// GetReadyz Readiness
	// (GET /readyz)
	GetReadyz(ctx context.Context, request GetReadyzRequestObject) (GetReadyzResponseObject, error)
	// GetStatus Service status (stub)
	// (GET /v1/status)
	GetStatus(ctx context.Context, request GetStatusRequestObject) (GetStatusResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetJWKS operation middleware
func (sh *strictHandler) GetJWKS(w http.ResponseWriter, r *http.Request, params GetJWKSParams) {
	var request GetJWKSRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetJWKS(ctx, request.(GetJWKSRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetJWKS")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetJWKSResponseObject); ok {
		if err := validResponse.VisitGetJWKSResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetHealthz operation middleware
func (sh *strictHandler) GetHealthz(w http.ResponseWriter, r *http.Request) {
	var request GetHealthzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealthz(ctx, request.(GetHealthzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealthz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthzResponseObject); ok {
		if err := validResponse.VisitGetHealthzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetReadyz operation middleware
func (sh *strictHandler) GetReadyz(w http.ResponseWriter, r *http.Request) {
	var request GetReadyzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetReadyz(ctx, request.(GetReadyzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetReadyz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetReadyzResponseObject); ok {
		if err := validResponse.VisitGetReadyzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetStatus operation middleware
func (sh *strictHandler) GetStatus(w http.ResponseWriter, r *http.Request) {
	var request GetStatusRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetStatus(ctx, request.(GetStatusRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetStatus")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetStatusResponseObject); ok {
		if err := validResponse.VisitGetStatusResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}